lsp: Move DocumentCache into extra file
It gets too big :-)
